PrepUploader/config/requirements.txt
pandas
requests
python-dotenv
numpy